[SPARK-34952][SQL][FOLLOWUP] Move aggregates to a separate package
