+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/expressions/aggregate/Min.java
@@ -15,36 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.connector.expressions;
+package org.apache.spark.sql.connector.expressions.aggregate;
 
 import org.apache.spark.annotation.Evolving;
+import org.apache.spark.sql.connector.expressions.FieldReference;
+import org.apache.spark.sql.connector.expressions.aggregate.AggregateFunc;
 
 /**
- * An aggregate function that returns the number of the specific row in a group.
+ * An aggregate function that returns the minimum value in a group.
  *
  * @since 3.2.0
  */
 @Evolving
-public final class Count implements AggregateFunc {
+public final class Min implements AggregateFunc {
   private final FieldReference column;
-  private final boolean isDistinct;
 
-  public Count(FieldReference column, boolean isDistinct) {
-    this.column = column;
-    this.isDistinct = isDistinct;
-  }
+  public Min(FieldReference column) { this.column = column; }
 
   public FieldReference column() { return column; }
-  public boolean isDistinct() { return isDistinct; }
 
   @Override
-  public String toString() {
-    if (isDistinct) {
-      return "COUNT(DISTINCT " + column.describe() + ")";
-    } else {
-      return "COUNT(" + column.describe() + ")";
-    }
-  }
+  public String toString() { return "MIN(" + column.describe() + ")"; }
 
   @Override
   public String describe() { return this.toString(); }
